+++ src/main/scala/org/jesperdj/scalaray/shape/DifferentialGeometry.scala
@@ -28,8 +28,8 @@ abstract class DifferentialGeometry {
 	val normal: Normal
 
 	// Surface parameter coordinates
-	val u: Double
-	val v: Double
+	val u: Float
+	val v: Float
 
 	// Partial derivatives of the surface position
 	val dpdu: Vector
